+++ src/Utils/FlashClasses.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Core;
+namespace App\Utils;
 
 class FlashClasses
 {
