+++ src/Utils/QuantityFormatter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Core;
+namespace App\Utils;
 
 /**
  * Utility class to format quantities.
